Use <CreateAccessCodeForm/> & <EditAccessCodeForm/> directly in <AccessCodeTable/>
