+++ src/lib/seam/components/AccessCodeDetails/AccessCodeDetails.element.ts
@@ -8,6 +8,7 @@ export const props: ElementProps<AccessCodeDetailsProps> = {
   accessCodeId: 'string',
   disableLockUnlock: 'boolean',
   onBack: 'object',
+  onEdit: 'object',
   className: 'string',
 }
